refactor file storage index and add tabs
